// src/gql/bootcampNeli.d.ts
export type BootcampNeliType = {
  id: string
  firstName: string
  lastName: string
  email: string
}

export type BootcampNeliResponse = {
  id: string
  firstName: string
  lastName: string
  email: string
}

export type BootcampNeliInput = {
  firstName: string
  lastName: string
  email: string
}

export type BootcampNeliFormData = {
  firstName: string
  lastName: string
  email: string
}